Updated numbers to floats

+++ server/models/buoydata.js
@@ -2,24 +2,24 @@
 module.exports = function(sequelize, DataTypes) {
   var buoydata = sequelize.define('buoydata', {
     BUOYID: DataTypes.INTEGER,
-    YY: DataTypes.NUMERIC,
-    MM: DataTypes.NUMERIC,
-    DD: DataTypes.NUMERIC,
-    hh: DataTypes.NUMERIC,
-    month: DataTypes.NUMERIC,
-    WDIR: DataTypes.NUMERIC,
-    WSPD: DataTypes.NUMERIC,
-    GST: DataTypes.NUMERIC,
-    WVHT: DataTypes.NUMERIC,
-    DPD: DataTypes.NUMERIC,
-    APD: DataTypes.NUMERIC,
-    MWD: DataTypes.NUMERIC,
-    PRES: DataTypes.NUMERIC,
-    ATMP: DataTypes.NUMERIC,
-    WTMP: DataTypes.NUMERIC,
-    DEWP: DataTypes.NUMERIC,
-    VIS: DataTypes.NUMERIC,
-    TIDE: DataTypes.NUMERIC
+    YY: DataTypes.FLOAT,
+    MM: DataTypes.FLOAT,
+    DD: DataTypes.FLOAT,
+    hh: DataTypes.FLOAT,
+    month: DataTypes.FLOAT,
+    WDIR: DataTypes.FLOAT,
+    WSPD: DataTypes.FLOAT,
+    GST: DataTypes.FLOAT,
+    WVHT: DataTypes.FLOAT,
+    DPD: DataTypes.FLOAT,
+    APD: DataTypes.FLOAT,
+    MWD: DataTypes.FLOAT,
+    PRES: DataTypes.FLOAT,
+    ATMP: DataTypes.FLOAT,
+    WTMP: DataTypes.FLOAT,
+    DEWP: DataTypes.FLOAT,
+    VIS: DataTypes.FLOAT,
+    TIDE: DataTypes.FLOAT
   }, {
     classMethods: {
       associate: function(models) {
